pyproject.toml
[project]
name = "job-search-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "muradeliyev", email = "muradaliyev2229@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "ddgs>=9.10.0",
    "langchain>=1.2.0",
    "langchain-community>=0.4.1",
    "langchain-google-genai>=4.1.2",
    "pdfplumber>=0.11.8",
    "pydantic-settings>=2.12.0",
    "httpx>=0.26.0",
    "beautifulsoup4>=4.12.3",
]

[project.scripts]
job-search-agent = "job_search_agent:main"

[build-system]
requires = ["uv_build>=0.9.13,<0.10.0"]
build-backend = "uv_build"
